ref: Fix linter error (#1981)

+++ Sources/Sentry/include/SentryCrashWrapper.h
@@ -30,7 +30,7 @@ SENTRY_NO_INIT
 
 - (NSDictionary *)systemInfo;
 
-- (NSInteger)freeMemory;
+- (uint64_t)freeMemory;
 
 @end
